requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
boto3==1.34.14
aiobotocore==2.9.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-dotenv==1.0.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
structlog==24.1.0
cryptography==41.0.7
python-jose==3.3.0
